pyproject.toml
[project]
name = "template-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.8.2",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pymysql>=1.1.2",
    "werkzeug>=3.1.3",
]
